E1-S6 admin command tree and E11-S3 lore renderer
/ascent now carries the PRD §6.4.1 admin tree: give <player> money|xp
(books, dust, scrolls and spawners answer "not yet" until their epics),
rank set|add <player> <n> clamped to the ladder, debug tps|db|items,
plus the existing version and reload. The whole tree sits behind
ascent.admin (LuckPerms grants it) and every action writes an
admin_actions row, asynchronously, with the arguments as JSON and the
nil UUID for the console. AdminService keeps the logic Bukkit-free so
it is unit tested against the in-memory doubles.

LoreRenderer is the single writer of custom item lore: enchant lines
<tier_color>Name Level highest tier first then by name, PROTECTED when
scrolled, then the kind footer; a [n] count on the display name when
there is at least one enchant. It keeps the item's original name in
persistent data and replaces the whole lore, so re-rendering is a
no-op. The pure parts are tested without a server.

Also fixes the Vault bridge compile error (a double balance passed to
a long parameter).

+++ ascent-plugin/src/main/java/gg/ascent/plugin/economy/VaultEconomy.java
@@ -325,7 +325,7 @@ public final class VaultEconomy implements Economy {
     return Bukkit.getOfflinePlayer(name);
   }
 
-  private static EconomyResponse success(long amount, long balance) {
+  private static EconomyResponse success(long amount, double balance) {
     return new EconomyResponse(amount, balance, ResponseType.SUCCESS, null);
   }
